TeleOp and Abstract Hardware Class
The abstract hardware class extends linear op mode but because its abstract it doesn't need to include the runOpMode() method. The TeleOp class extends the Hardware class and inherits all initialized hardware components and methods. This makes it so you don't have to initialize a hardware object and use dot notation, it just inherits everything.